requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
certifi==2023.5.7
chardet==5.1.0
charset-normalizer==3.1.0
crispy-bootstrap4==2022.1
Django==4.2.2
django-bootstrap-v5==1.0.11
django-crispy-forms==2.0
gunicorn==21.2.0
idna==3.4
packaging==23.1
Pillow==9.5.0
pytz==2023.3
requests==2.31.0
soupsieve==2.4.1
sqlparse==0.4.4
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.3
